TT#87953 [ec2/bootstrap.sh] Support -l option to list supported distributions
Nowadays ec2/bootstrap.sh is supposed to be invoked with the
distribution that should be created/updated. To avoid having
to manually identify which ones should be updated when building
a new base AMI, let's introduce `-l` option to list the supported
distributions.

This allows us to update the base systems and the build environments
running via:

  for distri in $(ec2/bootstrap.sh -l) ; do
    sudo ec2/bootstrap.sh -u "$distri"
  done

While at it improve stdout/stderr handling.

Change-Id: I9694ba6bdaf1019c04f1014b65a2f65e0f0e221a